# Conflicts:
#   include/mega/megaclient.h

+++ bindings/ios/include/MEGASdk.h
@@ -20,7 +20,6 @@
  */
 
 #import <Foundation/Foundation.h>
-#import <AssetsLibrary/AssetsLibrary.h>
 
 #import "MEGAAccountDetails.h"
 #import "MEGAAchievementsDetails.h"
@@ -498,7 +497,7 @@ typedef NS_ENUM(NSInteger, AdsFlag) {
 /**
  * @brief The total number of nodes in the account
  */
-@property (readonly, nonatomic) NSUInteger totalNodes;
+@property (readonly, nonatomic) unsigned long long totalNodes;
 
 /**
  * @brief The master key of the account.
